cloud-functions/remove/requirements.txt
google-cloud-datastore==1.9.0
google-cloud-secret-manager==0.2.0
sentry-sdk==0.14.2